// vitis-vision-front-end/src/app/core/api/models/vine-response.ts
/* tslint:disable */
/* eslint-disable */
export interface VineResponse {
  bolesCount?: number;
  createdAt?: string;
  createdBy?: string;
  formationType?: string;
  id?: number;
  lastUpdatedAt?: string;
  lastUpdatedBy?: string;
  plantingDate?: string;
  rowNumber?: number;
  variety?: string;
  vineNumber?: number;
}