// vitis-vision-front-end/src/app/core/api/models/device-response.ts
/* tslint:disable */
/* eslint-disable */
export interface DeviceResponse {
  active?: boolean;
  createdAt?: string;
  createdBy?: string;
  description?: string;
  deviceType?: string;
  id?: number;
  installationDate?: string;
  lastUpdatedAt?: string;
  lastUpdatedBy?: string;
  manufacturer?: string;
  name?: string;
}